+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Start some tests using the Cypress framework
 * Update php-cs-fixer to 3.8.0
 * Experimental compatibility with Lizmap 3.6
+* Support Extent input
 
 ## 0.1.1 - 2021-03-31